Consolidate `db::model` and `types`
