+++ bin/inx-chronicle/src/api/stardust/v2/routes.rs
@@ -8,9 +8,9 @@ use axum::{
     routing::*,
     Router,
 };
-use chronicle::{
-    db::MongoDb,
-    types::stardust::block::{BlockId, MilestoneId, OutputId, Payload, TransactionId},
+use chronicle::db::{
+    model::stardust::block::{BlockId, MilestoneId, OutputId, Payload, TransactionId},
+    MongoDb,
 };
 use futures::TryStreamExt;
 
@@ -51,7 +51,7 @@ pub fn routes() -> Router {
 }
 
 async fn block(database: Extension<MongoDb>, Path(block_id): Path<String>) -> ApiResult<BlockResponse> {
-    let block_id_dto = BlockId::from_str(&block_id).map_err(ParseError::StorageType)?;
+    let block_id_dto = BlockId::from_str(&block_id).map_err(ParseError::Model)?;
     let rec = database.get_block(&block_id_dto).await?.ok_or(ApiError::NoResults)?;
     Ok(BlockResponse {
         protocol_version: rec.inner.protocol_version,
@@ -62,7 +62,7 @@ async fn block(database: Extension<MongoDb>, Path(block_id): Path<String>) -> Ap
 }
 
 async fn block_raw(database: Extension<MongoDb>, Path(block_id): Path<String>) -> ApiResult<Vec<u8>> {
-    let block_id_dto = BlockId::from_str(&block_id).map_err(ParseError::StorageType)?;
+    let block_id_dto = BlockId::from_str(&block_id).map_err(ParseError::Model)?;
     let rec = database.get_block(&block_id_dto).await?.ok_or(ApiError::NoResults)?;
     Ok(rec.raw)
 }
@@ -71,7 +71,7 @@ async fn block_metadata(
     database: Extension<MongoDb>,
     Path(block_id): Path<String>,
 ) -> ApiResult<BlockMetadataResponse> {
-    let block_id_dto = BlockId::from_str(&block_id).map_err(ParseError::StorageType)?;
+    let block_id_dto = BlockId::from_str(&block_id).map_err(ParseError::Model)?;
     let rec = database.get_block(&block_id_dto).await?.ok_or(ApiError::NoResults)?;
 
     Ok(BlockMetadataResponse {
@@ -93,7 +93,7 @@ async fn block_children(
     Pagination { page_size, page }: Pagination,
     Expanded { expanded }: Expanded,
 ) -> ApiResult<BlockChildrenResponse> {
-    let block_id_dto = BlockId::from_str(&block_id).map_err(ParseError::StorageType)?;
+    let block_id_dto = BlockId::from_str(&block_id).map_err(ParseError::Model)?;
     let blocks = database
         .get_block_children(&block_id_dto, page_size, page)
         .await?
@@ -123,7 +123,7 @@ async fn block_children(
 }
 
 async fn output(database: Extension<MongoDb>, Path(output_id): Path<String>) -> ApiResult<OutputResponse> {
-    let output_id = OutputId::from_str(&output_id).map_err(ParseError::StorageType)?;
+    let output_id = OutputId::from_str(&output_id).map_err(ParseError::Model)?;
     let output_res = database
         .get_output(&output_id.transaction_id, output_id.index)
         .await?
@@ -169,7 +169,7 @@ async fn output_metadata(
     database: Extension<MongoDb>,
     Path(output_id): Path<String>,
 ) -> ApiResult<OutputMetadataResponse> {
-    let output_id = OutputId::from_str(&output_id).map_err(ParseError::StorageType)?;
+    let output_id = OutputId::from_str(&output_id).map_err(ParseError::Model)?;
     let output_res = database
         .get_output(&output_id.transaction_id, output_id.index)
         .await?
@@ -221,7 +221,7 @@ async fn transaction_included_block(
     database: Extension<MongoDb>,
     Path(transaction_id): Path<String>,
 ) -> ApiResult<BlockResponse> {
-    let transaction_id_dto = TransactionId::from_str(&transaction_id).map_err(ParseError::StorageType)?;
+    let transaction_id_dto = TransactionId::from_str(&transaction_id).map_err(ParseError::Model)?;
     let rec = database
         .get_block_for_transaction(&transaction_id_dto)
         .await?
@@ -236,7 +236,7 @@ async fn transaction_included_block(
 }
 
 async fn milestone(database: Extension<MongoDb>, Path(milestone_id): Path<String>) -> ApiResult<MilestoneResponse> {
-    let milestone_id_dto = MilestoneId::from_str(&milestone_id).map_err(ParseError::StorageType)?;
+    let milestone_id_dto = MilestoneId::from_str(&milestone_id).map_err(ParseError::Model)?;
     database
         .get_milestone_record(&milestone_id_dto)
         .await?
