+++ src/db/mod.rs
@@ -1,6 +1,8 @@
 // Copyright 2022 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
+/// Module containing the error type.
+pub mod error;
 /// Module containing database record models.
 pub mod model;
