+++ src/db/model/mod.rs
@@ -1,6 +1,10 @@
 // Copyright 2022 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
+#![allow(missing_docs)] // TODO Remove this once everything has settled.
+
+/// Module containing the ledger data models.
+pub mod ledger;
 /// Module containing Stardust data models.
 #[cfg(feature = "stardust")]
 pub mod stardust;
@@ -8,3 +12,5 @@ pub mod stardust;
 pub mod status;
 /// Module containing sync models.
 pub mod sync;
+/// Module contain utility functions.
+pub mod util;
