+++ src/db/model/stardust/block/block_id.rs
@@ -6,6 +6,8 @@ use std::str::FromStr;
 use bee_block_stardust as bee;
 use serde::{Deserialize, Serialize};
 
+use crate::db;
+
 #[derive(Clone, Debug, PartialEq, Serialize, Deserialize, Hash, Ord, PartialOrd, Eq)]
 #[serde(transparent)]
 pub struct BlockId(#[serde(with = "serde_bytes")] pub Box<[u8]>);
@@ -23,7 +25,7 @@ impl From<bee::BlockId> for BlockId {
 }
 
 impl TryFrom<BlockId> for bee::BlockId {
-    type Error = crate::types::error::Error;
+    type Error = db::error::Error;
 
     fn try_from(value: BlockId) -> Result<Self, Self::Error> {
         Ok(bee::BlockId::new(value.0.as_ref().try_into()?))
@@ -31,7 +33,7 @@ impl TryFrom<BlockId> for bee::BlockId {
 }
 
 impl FromStr for BlockId {
-    type Err = crate::types::error::ParseError;
+    type Err = db::error::Error;
 
     fn from_str(s: &str) -> Result<Self, Self::Err> {
         Ok(bee::BlockId::from_str(s)?.into())
