+++ src/db/model/stardust/block/output/basic.rs
@@ -1,73 +1,37 @@
 // Copyright 2022 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-use std::str::FromStr;
-
 use bee_block_stardust::output as bee;
 use serde::{Deserialize, Serialize};
 
 use super::{Feature, NativeToken, OutputAmount, UnlockCondition};
+use crate::db;
 
 #[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(transparent)]
-pub struct NftId(#[serde(with = "serde_bytes")] pub Box<[u8]>);
-
-impl NftId {
-    pub fn from_output_id_str(s: &str) -> Result<Self, crate::types::error::Error> {
-        Ok(bee::NftId::from(bee::OutputId::from_str(s)?).into())
-    }
+pub struct BasicOutput {
+    #[serde(with = "crate::db::model::util::stringify")]
+    pub amount: OutputAmount,
+    pub native_tokens: Box<[NativeToken]>,
+    pub unlock_conditions: Box<[UnlockCondition]>,
+    pub features: Box<[Feature]>,
 }
 
-impl From<bee::NftId> for NftId {
-    fn from(value: bee::NftId) -> Self {
-        Self(value.to_vec().into_boxed_slice())
-    }
-}
-
-impl TryFrom<NftId> for bee::NftId {
-    type Error = crate::types::error::Error;
-
-    fn try_from(value: NftId) -> Result<Self, Self::Error> {
-        Ok(bee::NftId::new(value.0.as_ref().try_into()?))
-    }
-}
-
-impl FromStr for NftId {
-    type Err = crate::types::error::ParseError;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        Ok(bee::NftId::from_str(s)?.into())
-    }
-}
-
-#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
-pub struct NftOutput {
-    amount: OutputAmount,
-    native_tokens: Box<[NativeToken]>,
-    nft_id: NftId,
-    unlock_conditions: Box<[UnlockCondition]>,
-    features: Box<[Feature]>,
-    immutable_features: Box<[Feature]>,
-}
-
-impl From<&bee::NftOutput> for NftOutput {
-    fn from(value: &bee::NftOutput) -> Self {
+impl From<&bee::BasicOutput> for BasicOutput {
+    fn from(value: &bee::BasicOutput) -> Self {
         Self {
             amount: value.amount(),
             native_tokens: value.native_tokens().iter().map(Into::into).collect(),
-            nft_id: (*value.nft_id()).into(),
             unlock_conditions: value.unlock_conditions().iter().map(Into::into).collect(),
             features: value.features().iter().map(Into::into).collect(),
-            immutable_features: value.immutable_features().iter().map(Into::into).collect(),
         }
     }
 }
 
-impl TryFrom<NftOutput> for bee::NftOutput {
-    type Error = crate::types::error::Error;
+impl TryFrom<BasicOutput> for bee::BasicOutput {
+    type Error = db::error::Error;
 
-    fn try_from(value: NftOutput) -> Result<Self, Self::Error> {
-        Ok(Self::build_with_amount(value.amount, value.nft_id.try_into()?)?
+    fn try_from(value: BasicOutput) -> Result<Self, Self::Error> {
+        Ok(Self::build_with_amount(value.amount)?
             .with_native_tokens(
                 Vec::from(value.native_tokens)
                     .into_iter()
@@ -86,12 +50,6 @@ impl TryFrom<NftOutput> for bee::NftOutput {
                     .map(TryInto::try_into)
                     .collect::<Result<Vec<_>, _>>()?,
             )
-            .with_immutable_features(
-                Vec::from(value.immutable_features)
-                    .into_iter()
-                    .map(TryInto::try_into)
-                    .collect::<Result<Vec<_>, _>>()?,
-            )
             .finish()?)
     }
 }
@@ -101,8 +59,8 @@ pub(crate) mod test {
     use mongodb::bson::{from_bson, to_bson};
 
     use super::*;
-    use crate::types::stardust::block::output::{
-        feature::test::{get_test_issuer_block, get_test_metadata_block, get_test_sender_block, get_test_tag_block},
+    use crate::db::model::stardust::block::output::{
+        feature::test::{get_test_metadata_block, get_test_sender_block, get_test_tag_block},
         native_token::test::get_test_native_token,
         unlock_condition::test::{
             get_test_address_condition, get_test_expiration_condition, get_test_storage_deposit_return_condition,
@@ -111,26 +69,15 @@ pub(crate) mod test {
     };
 
     #[test]
-    fn test_nft_id_bson() {
-        let nft_id = NftId::from(rand_nft_id());
-        let bson = to_bson(&nft_id).unwrap();
-        assert_eq!(nft_id, from_bson::<NftId>(bson).unwrap());
-    }
-
-    #[test]
-    fn test_nft_output_bson() {
-        let output = get_test_nft_output();
+    fn test_basic_output_bson() {
+        let output = get_test_basic_output();
         let bson = to_bson(&output).unwrap();
-        assert_eq!(output, from_bson::<NftOutput>(bson).unwrap());
+        assert_eq!(output, from_bson::<BasicOutput>(bson).unwrap());
     }
 
-    pub(crate) fn rand_nft_id() -> bee::NftId {
-        bee_test::rand::bytes::rand_bytes_array().into()
-    }
-
-    pub(crate) fn get_test_nft_output() -> NftOutput {
-        NftOutput::from(
-            &bee::NftOutput::build_with_amount(100, rand_nft_id())
+    pub(crate) fn get_test_basic_output() -> BasicOutput {
+        BasicOutput::from(
+            &bee::BasicOutput::build_with_amount(100)
                 .unwrap()
                 .with_native_tokens(vec![get_test_native_token().try_into().unwrap()])
                 .with_unlock_conditions(vec![
@@ -151,12 +98,6 @@ pub(crate) mod test {
                         .unwrap(),
                     get_test_metadata_block().try_into().unwrap(),
                     get_test_tag_block().try_into().unwrap(),
-                ])
-                .with_immutable_features(vec![
-                    get_test_issuer_block(bee_test::rand::address::rand_address().into())
-                        .try_into()
-                        .unwrap(),
-                    get_test_metadata_block().try_into().unwrap(),
                 ])
                 .finish()
                 .unwrap(),
