--- src/types/stardust/mod.rs
@@ -1,4 +1,0 @@
-// Copyright 2022 IOTA Stiftung
-// SPDX-License-Identifier: Apache-2.0
-
-pub mod block;
